+++ R/check-rows.R
@@ -2,6 +2,8 @@
 #'
 #' Checks the number of rows in data.
 #'
+#' By default (\code{min_row = 1}) data must include at least one row.
+#'
 #' @inheritParams check_data_frame
 #' @param min_row A count of the minimum number of rows.
 #' @param max_row A count of the maximum number of rows.
@@ -10,12 +12,12 @@
 #' data.
 #' @seealso \code{\link{datacheckr}}
 #' @export
-check_rows <- function(data, min_row = 0, max_row = max_nrow(), data_name = substitute(data)) {
+check_rows <- function(data, min_row = 1, max_row = max_nrow(), data_name = substitute(data)) {
   data_name <- as.character(data_name)
   data <- check_data_frame(data, data_name = data_name)
 
-  if (!is_count(min_row)) check_stop("min_row must be a count")
-  if (!is_count(max_row)) check_stop("max_row must be a count")
+  check_count(min_row)
+  check_count(max_row)
 
   if (min_row < 0) check_stop("min_row must not be less than 0")
   if (max_row < min_row) check_stop("max_row must not be less than min_row")