+++ R/utils.R
@@ -10,8 +10,14 @@ is_POSIXct <- function(x) inherits(x, "POSIXct")
 
 check_stop <- function(...) stop(..., call. = FALSE)
 
-check_data_name <- function(data_name)
-  if (!is_string(data_name)) check_stop("data_name must be a string")
+check_string <- function(x)
+  if (!is_string(x)) check_stop(substitute(x), " must be a string")
+
+check_flag <- function(x)
+  if (!is_flag(x)) check_stop(substitute(x), " must be a flag")
+
+check_count <- function(x)
+  if (!is_count(x)) check_stop(substitute(x), " must be a count")
 
 plural <- function(x, n = 1, end = "") paste0(x, ifelse(n != 1, "s", ""), end)
 isare <- function(n) ifelse(n > 1, "are", "is")