+++ tests/testthat/test-check-rows.R
@@ -19,9 +19,10 @@ test_that("check_rows checks rows", {
   x <- data.frame(y = 2)
   expect_identical(check_rows(x), x)
   expect_error(check_rows(x, min_row = 2), "x must have at least 2 rows")
-  expect_error(check_rows(x, max_row = 0), "x must have no more than 0 rows")
+  expect_error(check_rows(x, max_row = 0), "max_row must not be less than min_row")
+  expect_error(check_rows(x, min_row = 0, max_row = 0), "x must have no more than 0 rows")
   x <- x[FALSE,,drop = FALSE]
-  expect_identical(check_rows(x), x)
-  expect_error(check_rows(x, min_row = 1), "x must have at least 1 row")
+  expect_identical(check_rows(x, min_row = 0), x)
+  expect_error(check_rows(x), "x must have at least 1 row")
+  expect_error(check_rows(x, min_row = 2), "x must have at least 2 rows")
 })
-